+++ spec/gen/ts/index.ts
@@ -7,6 +7,7 @@ export * from "./recipe.js";
 export * from "./relationship.js";
 export * from "./setting-value.js";
 export * from "./issue.js";
+export * from "./artifact.js";
 export * from "./reply-plan.js";
 export * from "./model-capabilities.js";
 export * from "./entity.js";